Fixed pluginfile (#31)
Due to a wrong context in lib.php (pluginfile), probably as a result of copy-paste or plugin rename, the authentication method logo did always return a file not found error. That is most likely the cause for issue #30 which I could reproduce on my local setup (Moodle Version 3.9). I propose to change that context to the plugin name to achieve the expected behavior.